Clean up transformers (#12)
* Split the protocols

* Renamed result type

* Moved type to its file
