+++ Sources/URLRequestable/DataTypes.swift
@@ -1,5 +1,5 @@
 //
-//  Defines.swift
+//  DataTypes.swift
 //
 //  Created by Waqar Malik on 4/27/23.
 //
@@ -10,7 +10,3 @@ public typealias DecodableHandler<T: Decodable> = (Result<T, Error>) -> Void
 public typealias SerializableHandler = (Result<Any, Error>) -> Void
 public typealias DataHandler<T> = (Result<T, Error>) -> Void
 public typealias ErrorHandler = (Error?) -> Void
-public typealias Transformer<InputType, OutputType> = (InputType) throws -> OutputType
-
-@available(iOS 15, tvOS 15, watchOS 8, macCatalyst 15, macOS 12, *)
-public typealias AsyncTransformer<InputType, OutputType> = (InputType) async throws -> OutputType
