+++ Sources/URLRequestable/URLRequestable.swift
@@ -8,11 +8,12 @@ import Foundation
 import HTTPTypes
 
 public typealias URLDataResponse = (data: Data, response: URLResponse)
+public typealias Transformer<InputType, OutputType> = (InputType) throws -> OutputType
 
 public protocol URLRequestable {
-    associatedtype Response
+    associatedtype ResultType
 
-    typealias ResponseTransformer = Transformer<URLDataResponse, Response>
+    typealias URLResponseTransformer = Transformer<URLDataResponse, ResultType>
 
     var apiBaseURLString: String { get }
     var method: URLRequest.Method { get }
@@ -21,7 +22,7 @@ public protocol URLRequestable {
     var body: Data? { get }
     var queryItems: [URLQueryItem]? { get }
 
-    var transformer: ResponseTransformer { get }
+    var transformer: URLResponseTransformer { get }
 
     func url(queryItems: [URLQueryItem]?) throws -> URL
     func urlRequest(headers: [HTTPHeader]?, queryItems: [URLQueryItem]?) throws -> URLRequest
@@ -70,8 +71,8 @@ public extension URLRequestable {
     }
 }
 
-public extension URLRequestable where Response: Decodable {
-    var transformer: ResponseTransformer {
+public extension URLRequestable where ResultType: Decodable {
+    var transformer: URLResponseTransformer {
         JSONDecoder.transformer()
     }
 }
